Move TaskFile and TaskIdent to submodules.
