+++ src/scheduler/task_file.rs
@@ -1,0 +1,33 @@
+use log::debug;
+use std::fs::{self, create_dir_all, remove_file, File};
+use std::hash::{Hash, Hasher};
+use std::io::Error;
+use std::path::PathBuf;
+
+#[derive(Debug)]
+pub(crate) struct TaskFile {
+    pub(crate) file: File,
+    pub(crate) path: PathBuf,
+}
+
+impl PartialEq for TaskFile {
+    fn eq(&self, other: &Self) -> bool {
+        self.path.eq(&other.path)
+    }
+}
+
+impl Eq for TaskFile {}
+
+impl std::hash::Hash for TaskFile {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.path.hash(state)
+    }
+}
+
+impl TaskFile {
+    pub(crate) fn destroy(&self) -> Result<(), Error> {
+        remove_file(self.path.clone())?;
+        debug!("Removing TaskFile from queue");
+        Ok(())
+    }
+}
